the form is extracted out with an update to validatePropert function
