+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tech-startup",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
